Use afterCreate with autorun on rootmodel to do localstorage and url testing
